Compiler: fix bug/regression with incremental .NET builds
Attempt to lazy-compile when encountering functions that aren't compiled in
ShaderProcessor. This can happen when an up-to-date version of a function's
assembly can be loaded from disk.

Before #282 this bug triggered only on methods pre-compiled in Uno.Runtime.Core.dll,
and after the bug started triggering on any method from another package.

This PR should fix the bug for all methods.