UnoDoc: handle arrays of arrays
It seems this generator is written without concern for arrays of
arrays, which we use in our API.

According to the docs-team, this was ad-hoc hacked around when
outputting the API in the past. But let's instead do this correctly,
by peeling away multiple arrays rather than just a single one.

This prevents an error about missing namespace for `byte[]` when
updating API docs.