rename Library/Core/ to lib/
This is mainly because it's much shorter to type, and has only one path
component instead of two. It will save a lot of time in the long run when
doing work from the terminal.

This patch touches very many files, but the change is much less overwhelming
than it seems. Rebasing is trivial as long as no new files are added to the
tree. We don't see many new files these days, so I don't really think this
change will be very inconvenient for anyone.