UnoCore: remove try..finally blocks in GL
The underlying low-level implementations don't throw, so these clauses are
pretty much a waste of CPU-cycles and confusing to the reader. :)

Even if the low-level code did throw, that would be considered a bug on their
end and not something we need to worry about in here.

We also don't try..finally in the other methods, so let's be consistent.