C++: implement GetEntity() in CppDecompiler
If passed a method, the decompiler will now expand to the qualified name of
the generated C++ method.

This is useful to register callbacks in C APIs - as long as the parameters
are matching, and the method is static, the generated method can be used
as callback directly. For example, when working with GStreamer:

    G_CALLBACK(@{OnPadAdded(GstElementPtr,GstPadPtr,GstPlayer)})