UnoCore: rename Obj-C file to fix iOS build error
When using GStreamer.framework in an iOS project, the build fails with linker
errors caused by missing symbols from Support.cpp - uLog, uFatal and so on.

Renaming Support.mm to SupportApple.mm fixes the issue.

Strangely, when not using GStreamer.framework the build passes without
renaming the file, and this has not been a problem before.