Android: remove gnustl packaging option
This was a hint for the build system to choose the first one in case multiple
versions of the library are found. Since we no longer use gnustl after it was
removed in NDK r18 (f99047247218c30cae5f8cb5c8d437b50a45c151), we no longer
need the hint (this was mentioned in #109).