Android: add google repository to all projects #130
Myself and others have occasionally experienced trouble when downloading
dependencies when building for Android, and this seems to fix the problem.

I haven't seen this problem in a while and it seems to have been caused by
temporary issues on Google servers, but it doesn't cause any harm to add the
fix anyway, hopefully avoiding similar trouble in the future.