Android: support both Landscape modes
Currently, setting orientation to Landscape will lock the app in only one
landscape mode and the app will not rotate according to the sensor. This is
not the expected behaviour.

This fixes so that setting orientation to Landscape will rotate the app
according to the sensor. It is also possible to specify LandscapeLeft or
LandscapeRight to force only one of the landscape modes.