include global android-build-tools config
android-build-tools is an NPM package that automatically installs Android
SDK and NDK, required to build Android apps, and generates a config file for
Uno, similar to 'fuse install android'.

This package can be installed globally like this

    npm install -g android-build-tools

The good thing about this is that we no longer need to install Fuse Studio
to get automatic access to the Android SDK and NDK which is especially
convenient when setting up CI for Fuse apps.