FuseJS.Transpiler: upgrade cached-path-relative
Due to CVE-2018-16472, it's recommended to upgrade this package. As
the *only* change betweem 1.0.1 and 1.0.2 is the fix for this
vulnerability, let's upgrade it.

The vulnerability doesn't seem to be a big deal to our users, as they
control the scripts that run in their apps themselves. So there's no
reason to back-port this to older releases.