AdbRunner: introduce Android.IgnoreNetworkDevices config #206
Some users need to run apps on network devices, but v1.12 introduced a change
that ignores network devices.

This patch makes ignoring network devices opt-in, and users that hasn't set
Android.IgnoreNetworkDevices will get the old behaviour restored, as things
used to be in pre-v1.12.

Users that want to ignore network devices can add this to their ~/.unoconfig:

    Android.IgnoreNetworkDevices: true