pack.sh: simplify version info
This drops $BRANCH and $SUFFIX, and simplifies logic (#208).

When updating an NPM package, the version field must be updated in
package.json, and all changes to package.json should be committed in Git.

This means that publishing an NPM package, pre-release or not, will always
involve updating package.json, committing, and pushing a tag in Git.

This also means that pack.sh should not manipulate version info read from
package.json, as it used to do in pre-NPM times for legacy reasons.