Uno.ProjectFormat: use $npm_package_version as default version
This will use the version number specified in package.json, if available,
when Uno is invoked via a npm-script, unless an explicit version number was
specified in the project file or on the command-line.

With this, we no longer need Bash to forward the version number from
package.json onto Uno in a cross-platform way. This should make it a little
bit easier to maintain Uno packages using NPM.

See package.json in Fuselibs for an example.