app-mac: fix Xcode 10 regression
Apparently all OpenGL apps powered by Mono on Mac stopped rendering if
built on a machine with Xcode 10 (or newer) installed.

Xcode 9.4 on macOS High Sierra is the last known to work configuration,
and apps built on this configuration would still work fine on newer
macOS and Xcode, it sounds like.

This fixes OpenGL rendering in apps built with Xcode 10 (macOS Mojave)
and newer.

https://github.com/xamarin/xamarin-macios/issues/4959