make configuration directory lower-case
This makes the configuration part of the build output directory
lower-case, now making all path components of the default build
output directory lower-case.

Some users would argue that all lower-case directories appear more
modern, and look more familiar to users coming from other tooling.
Also, it's not a bad thing to update one's appearance once in a while,
and since we're preparing our next major release this seems like good
timing.

Please note that this is a breaking change (on case-sensitive
file-systems), though unlikely to affect users in practice (since
the majority of our audience are likely users of case-insensitive
file-systems, and it is rarely necessary to deal with your output
directory directly).