iOS: rename project property
We want to avoid using `-` in property names because it makes the name an
invalid identifier, and we want consistency with other property names.
Some properties are mapped in JavaScript and other languages.

The old name still works but with a deprecation warning asking you to change
the name if you are using this property.