iOS: drop legacy launch images
Legacy launch images are redundant beginning with iOS 12, as in they are
no longer used by iOS 12 and newer.

The existence of these launch images produces build warnings on some Xcode
versions, and because iOS 12 is fairly old at this point, and because we
have another way to specify a launch screen image, it seems like a good
idea to remove them.