remove VersionDirectory from libraries
Before when building libraries, output files were placed inside a
directory named after the current version of the library, e.g.

    lib/build/UnoCore/1.2.3/

This is a relic from a time where it was possible to install multiple
versions of the same library side by side using Uno Package Manager.

Uno Package Manager was removed in favor of NPM in 2019, so we no longer
have this requirement and can remove the last directory part, e.g.

    lib/build/UnoCore/

We should still be able to load libraries built by older versions of
Uno using the version directory.