unoproj: support lowerCamelCase property names
We want to accept property names written in lowerCamelCase, in unoproj-
files, unoconfig-files, and uxl-files; to better match JavaScript and
modern .NET, and best practices.

In other words, users can now chose if they want to keep their existing
UpperCamelCase style in project files or switch to lowerCamelCase, which
is recommended for new projects built on version 3.0.

We will update existing project and config files in upcoming patches, so
this comes with functional changes to begin with, making sure both
styles are tested.