macOS: prebuilt app loader
This adds a prebuilt app loader for macOS, downloaded via NuGet.

The binaries that come out on macOS are quite large and don't change
often, so move them to NuGet to avoid bloating the main Uno package.

Native binaries are also CPU specific (x64), so we'll publish another
set for arm64 at a later point.

* Configure assemblies via Info.plist and load via .NET reflection
  instead of editing assemblies, since edited macOS assemblies results
  in crashes on .NET 6.0.

* Move relevant source code to a new repository at
  https://github.com/fuse-open/mac-loader.