engine: switch to HttpClient
This switches to use HttpClient on .NET 6 because old WebClient and
friends are deprecated, silencing a few compile time warnings.

Note that HttpClient (unlike WebClient) should preferably be created
only once and kept alive through the lifetime of an application, in
order to not exhaust the number of sockets available under heavy loads.

Also it appears that enabling TLS 1.2 explicit is no longer necessary.