compiler: replace System.Drawing on macOS and Linux
This switches to use the 'sips' command on macOS or the 'convert'
command (from ImageMagick) on Linux to resize images/icons, since
System.Drawing isn't supported on those platforms. Doing this silences
the remaining compile-time warnings when building on .NET 6.

We have also tried integrating ImageMagick.NET but that adds 90MB of
native binaries to our package, seeming a bit excessive, so let's see
how far we get with the system commands first.