compiler: return "if-test and throw" when compiling 'assert'
'assert' statements were turned into method calls to a method that was
removed in 992b515, resulting in compile errors.

'assert' is still used in a few places, and this commit turns them into
"if-test and throw" in debug builds (noop in release builds) to fix the
compile errors.

    assert VALUE;

becomes when compiled:

    if (IS_DEBUG_BUILD && !VALUE)
        throw new Uno.InvalidOperationException(
            "Assertion failed: VALUE"
        );