remove xamarin-mac dependency
Xamarin.Mac isn't supported on .NET 6 (throws exception), blocking us
from building iOS apps, so this commit removes the dependency.

Xamarin.Mac was used to access Xcode Developer Teams, so we need a way
to replace this functionality. I have searched for a node module or a
CLI that can replace it, without luck, so it looks like we'll need to
implement a new tool to access Xcode Developer Teams at some point.