engine: lazy compilation for .NET targets
Pass -DLAZY to "uno build" to enable lazy compilation (experimental).

We've seen around 10x[1] speed-up on subsequent builds of a normal app
when lazy compilation is enabled, since there is minimal work to do
when most of the code is already built.

The feature is experimental for now, but we might want to take
advantage of this (e.g. in the Fuse X editor) when the feature
matures a bit more, to speed up loading times in a significant way
by prebuilding standard libraries (including Fuselibs), and better
caching.

1: More or less, depending on project size